relinktextoperators do quick elimination now
Instead of it being done by the text relinkoperator. This is because there
may exist tiddler types where just because the fromTitle doesn't show up
anywhere in the text, doesn't mean relinking shouldn't occur.

For instance, if fromTitle is "Jesse's cats". It might show up in a
javascript tiddler as "Jesse\'s cats", which isn't the same, but still
needs relinking.